requirements.txt
requests==2.32.3
python-dotenv==1.1.0
google-cloud-bigquery==3.41.0
streamlit==1.45.1